Quotes refactored + got voting to work
There is now block_quote.html which is not a full page, but just a
block of html that can be included in other pages. The voting
javascript is now in its own file. The single quote page and the quote
page page now use this new building block and the ajax voting style.